Made hcsr04::begin() set baseline value. Changed hcsr04::readRelativeToBaseline() to int16_t to ensure the same decimal precision as WaterPressure::readWaterLevel().

+++ embedded-systems/lib/hcsr04/hcsr04.h
@@ -20,7 +20,7 @@ namespace hcsr04
     float readDistance(); // returns distance in cm
     void calibrateZero(float known_level_cm);
     void setBaselineFromCurrentReading();
-    float readRelativeToBaseline();
+    int16_t readRelativeToBaseline();
     void setBaseline(float cm);
     unsigned long simulateEchoDurationFromCM(float cm); // used only in mock mode
 }
